+++ config.js
@@ -4,6 +4,11 @@ const configuracion = {
     version: "1.0.0",
     fechaCreacion: "2025-11-14",
     miembros: [
+		{
+			nombre: "Alexis Sevilla Serrano",
+			rol: "Desarrollador de Funcionalidades Web",
+			github: "Alexis(https://github.com/Alexis15014)"
+		},
         {
             nombre: "Kevin Alejandro Almeida Camacho",
             rol:"Diseñador Web",